+++ api/v1/public/share.php
@@ -35,6 +35,7 @@ ErrorResponse::ok([
     'expose_detail'    => (bool) $link['expose_detail'],
     'expose_map'       => (bool) $link['expose_map'],
     'expose_stats'     => (bool) $link['expose_stats'],
+    'expose_review_summary' => (bool) $link['expose_review_summary'],
     'requires_password'=> $needsPassword,
     'unlocked'         => $unlocked,
     'last_synced_at'   => $link['last_synced_at'],